players_new_v2/2020s_project/race_pipeline/requirements.txt
httpx>=0.27,<1.0
trafilatura>=1.12,<2.0
spacy>=3.7,<4.0
anthropic>=0.34,<1.0
python-dotenv>=1.0,<2.0
pandas>=2.2,<3.0
